examples/python/openai_sdk_1.x/requirements.txt
openai>=1.3.3, <3.0.0
python_dotenv>=1.0.0, <2.0.0
azure-ai-projects>=1.0.0
azure-identity>=1.19.0
httpx
tenacity
termcolor
black
ruff
